requirements.txt
aniso8601==4.1.0
appnope==0.1.0
asgiref==3.3.1
asn1crypto==0.24.0
attrs==18.2.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
cycler==0.10.0
decorator==4.3.2
Django==3.0.5
django-appconf==1.0.3
django-imagekit==4.0.2
djangorestframework==3.12.2
djongo==1.3.3
docutils==0.16
drf-yasg==1.20.0
entrypoints==0.3
et-xmlfile==1.0.1
gevent==20.6.2
greenlet==0.4.16
httplib2==0.17.0
idna==2.8
importlib-metadata==1.6.0
inflection==0.5.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.13.2
jsonschema==3.0.0a3
keyring==21.2.0
kiwisolver==1.1.0
limits==1.3
lxml==4.3.1
MarkupSafe==1.1.0
matplotlib==3.2.0
mistune==0.8.4
mpmath==1.1.0
multitasking==0.0.9
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.4
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.8
pandocfilters==1.4.2
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.0.0
pkginfo==1.5.0.1
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.5.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.9
Pygments==2.6.1
pymongo==3.11.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.14.10
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.9
pyzmq==17.1.2
qtconsole==4.4.3
readme-renderer==26.0
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
seaborn==0.10.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.12.0
soupsieve==1.7.1
SQLAlchemy==1.2.17
sqlparse==0.2.4
sympy==1.5.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
twine==3.1.1
tzlocal==1.5.1
uritemplate==3.0.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
